Integrate pytest instead of nosetests
