+++ setup.cfg
@@ -2,3 +2,7 @@
 max-line-length=120
 exclude=env,venv,node_modules
 
+[tool:pytest]
+testpaths=tests
+confcutdir=tests
+addopts=--tb=short
